Added perl program I wrote and forgot to track
Added run_fastq-dump_parallel.pl, a perl program I wrote and have tweaked over the past month or so. I appear to have forgotten to track it on Git- whoops! This program is straightforward- it uses GNU Parallel to run multiple instances of fastq-dump, a tool that downloads .fastq files from NCBI's SRA.