Add user-specified output folder. Move Bakta annotation to Bakta subworkflow
Apart from above: Have the process that downloads a Bakta database output the path to the database. This has no real function, but forces the Bakta annotation process to wait on the download, since it 'needs' the output the download process produces. Essentially just a way to get Nextflow to wait for the download the finish