Add option for file specifying mandatory regions for integrations. Adjust whitespace
Add parse_arg option for a path to a .tsv file containing viral names, start coordinates, and end coordinates. Any matches to specified viruses that do not overlap at all with the region on the reference viral genome's mandatory region will be discarded. Also, adjust whitespace to make clear which lines of code comments are associated with