Initial commit for change_contig_ids.py
This Python program is supposed to have two modes: rename, which parses an input .fasta format file and renames its contig IDs (the portion of the header line between > and the first whitespace) and revert, which takes in the output of some program with the replaced contig IDs and swaps them out for the original IDs. This should allow VIBES to maintain the same contig IDs (even when some programs have issues with them), allowing for output visual generatation.